dev/WP/feature#2 - created and organization pastes and implementation modular
