+++ lib/src/modules/authentication/authentication_module.dart
@@ -1,0 +1,10 @@
+import 'package:flutter_modular/flutter_modular.dart';
+
+import 'ui/pages/authentication_page.dart';
+
+class AuthenticationModule extends Module {
+  @override
+  void routes(RouteManager r) {
+    r.child('/', child: (context) => const AuthenticationPage());
+  }
+}
