+++ lib/src/modules/authentication/ui/pages/authentication_page.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class AuthenticationPage extends StatelessWidget {
+  const AuthenticationPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
