+++ pubspec.yaml
@@ -31,6 +31,7 @@ dependencies:
   flutter:
     sdk: flutter
   get_it: ^7.7.0
+  flutter_modular: ^6.3.0
 
 
   # The following adds the Cupertino Icons font to your application.
